Generate semantic conventions according to specification latest tagged version
This modifies the generation tool to generate the content from the latest tagged
version of the specification (previously it was just using whatever the spec repo's
working directory contained).

The generated files will be now placed in a sub-directory with a name matching the
tagged version number. The generated package will be placed under the "semconv" directory.

Note: we will use the new logic generate the semconv/v1.3.0 package in a separate future PR.