Add missing apply_to_spans section support
This bug was uncovered during spec review here:
https://github.com/open-telemetry/opentelemetry-specification/pull/2008#discussion_r731359448

apply_to_spans section was missing although it is implied to
be present in the description of how span changes section
is supposed to work.

This fixes the omission by adding apply_to_spans section.
This is not a new feature, just fixing a bug in file format description
and in the corresponding implementation.