Avoid overriding explicitly set grpc dial options
The contract of oteltracegrpc.WithDialOption is that the options
supplied become the tail end of the dial option slice that is eventually
supplied to grpc.Dial however in some cases the implementation violates
that contract and appends additional dial options at the end of said
slice.

This commit changes otlpconfig option processing to ensue that the
contract of oteltracegrpc.WithDialOption is upheld. That is, the options
supplied become the suffix of the slice used to dial.

Fixes #2940