BAD: Add registry of set data
This is a failed experiment that tried to hold set data in its own
registry and use a uint64 reference number to act as a fast map key and
underlie a set.

It failed because a Set is returned from all the NewSet* functions and
since it is not allocated to the heap there is no way to correctly set a
finalizer for the object.