Solve the problem with unsafe.Pointer
This is still a failed solution. It converts a pointer to a uintptr via
unsafe.Pointer, but it also then converts it back. This is going to
cause issues because the original pointer address is not guaranteed to
remain the same and casting back could end up pointing at invalid
memory.