Change NewLoggerConfig signature
Have the NewLoggerConfig function accept a slice instead of variadic
argument. This is going to be called, outside of direct testing, from
the LoggerProvider implementations when their Logger method is called.
That method already accepts a variadic argument meaning they will have a
slice to pass. This change avoids them having to do the following:

    c := NewLoggerConfig(options...)

Instead, they can just ...

    c := NewLoggerConfig(options)