Export the Instrument IsEmpty method (#5431)
From
https://github.com/open-telemetry/opentelemetry-go-contrib/pull/5654#discussion_r1617971674

Constructing a view requires an `Instrument` to be constructed. Given
`NewView` will not return an error directly when an empty instrument is
passed, it may be ideal for users to check this prior to making the
call. Instead of having all use-cases copy this code, export it so they
can just call the method.