Enable all benchmarks in CI (#5644)
Part of #4537

Here is the recent benchmark result from the CI:
-
https://github.com/open-telemetry/opentelemetry-go/actions/runs/10067601185
-
https://github.com/open-telemetry/opentelemetry-go/actions/runs/10068023613
-
https://github.com/open-telemetry/opentelemetry-go/actions/runs/10072452529

Though the results are pretty stable, they only run for very limited
benchmarks, which cannot reflect the stability when running more
benchmarks.

---

This PR enables all benchmarks in CI so we can obtain accurate stability
results. I also removed the `timeout` flag for the benchmark test, as
some benchmarks contain a lot of sub-benchmarks, which will timeout
anyway if we have a 60-second timeout. (Go does not offer timeout for
single benchmark or test)