Use self hosted runner to run benchmark (#5695)
Part of #4537

These are some benchmark results from the `actuated` runner. It is
pretty clear to me that this runner does not have a stable environment
to produce stable benchmark results:
-
https://github.com/open-telemetry/opentelemetry-go/actions/runs/10278899889
-
https://github.com/open-telemetry/opentelemetry-go/actions/runs/10275384692

Thus, I am switching the runner to the self-hosted runner, which uses
bare metal, to run benchmarks.

This is the request to use this type of runner.
https://github.com/open-telemetry/community/issues/2266

The underlying machine is m3.small.x86. It has 8 cores 64 GiB.
https://deploy.equinix.com/product/servers/m3-small/

---

This is an example of using this runner from otel java.
https://github.com/open-telemetry/opentelemetry-java/actions/runs/10277337397/job/28439208528

---

This runner cannot be triggered on a forked repo:
https://github.com/XSAM/opentelemetry-go/commit/faab7b90979749b5f01fb0dd9522bbcfaedd5854.
So, it is quite safe if we only let it run on the main branch.