sdk/trace: fix goroutine leak in ForceFlush on context cancel
When `batchSpanProcessor.ForceFlush` was called and the context was
canceled, the method's select statement returned the context err right
away.

This also meant that the `wait` chan was no longer being read from,
which meant that the goroutine attempting to write to it ended up
blocking and leaking forever.

The fix is to just make the `wait` chan buffered with a size of 1. It
can always be written to and will just be gc'd if the `ForceFlush`
method has already exited.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>