Ignore deprecation warnings about NameValidationScheme (#6449)
In 0.63.0, prometheus deprecates the `model.NameValidationScheme`
attribute.

https://github.com/open-telemetry/opentelemetry-go/actions/runs/13885046847/job/38848614767?pr=6442

> Error: exporter_test.go:478:6: SA1019: model.NameValidationScheme is
deprecated: This variable should not be used and might be removed in the
far future. If you wish to stick to the legacy name validation use
`IsValidLegacyMetricName()` and `LabelName.IsValidLegacy()` methods
instead. This variable is here as an escape hatch for emergency cases,
given the recent change from `LegacyValidation` to `UTF8Validation`,
e.g., to delay UTF-8 migrations in time or aid in debugging unforeseen
results of the change. In such a case, a temporary assignment to
`LegacyValidation` value in the `init()` function in your main.go or so,
could be considered. (staticcheck)

We probably want to keep supporting this scheme in the near future
though. So this ignores those deprecations, so we can upgrade the
package.

Unblocks https://github.com/open-telemetry/opentelemetry-go/pull/6442