attribute: fix JSON marshaling of Set
The Set type implemented json.Marshaler on a pointer receiver. Alas, an
attribute.Set is sometimes used as a value in otel packages. Most
notably, the NewSet function returns a value not a pointer.

Implementing MarshalJSON on a value receiver, like MarshalLog, enables
proper marshalling when a Set is used as a value that is not
addressable.