Update the required approvals policy
Based on SIG meeting discussions today: relax requirement for two
approvals from Approvers at different companies. Instead, require at
least one approver or the author to not work at the same company.

This means that if the author works at a different company than two
Approvers who have approved and work at the same company, the PR is
ready to merge.

This continues the enforcement that if an author and all Approvers
approving a PR work at the same company, additional qualified approvals
are needed.