docs: Add Go Code Review Comments and allow the use of testify (#6984)
Inspired by
https://github.com/open-telemetry/opentelemetry-go-contrib/pull/7522#discussion_r2190580677

I hope it will help new contributors. Also I want to codify that we are
OK when one uses `testify`, but I do not also want to force it (e.g.
https://github.com/open-telemetry/opentelemetry-go-contrib/pull/7522 was
not using it).

I think that as long as we use the standard library, `testify`, `go-cmp`
we are fine (and should not accept any other libraries than those).