Release exporters/prometheus/v0.59.1 (#7056)
Patch release for
https://github.com/open-telemetry/opentelemetry-go/issues/7039

### Changed

- Retract `v0.59.0` release of
`go.opentelemetry.io/otel/exporters/prometheus` module which appends
incorrect unit suffixes. (#7046)
- Change `go.opentelemetry.io/otel/exporters/prometheus` to no longer
deduplicate suffixes when UTF8 is enabled.
It is recommended to disable unit and counter suffixes in the exporter,
and manually add suffixes if you rely on the existing behavior. (#7044)

### Fixed

- Fix `go.opentelemetry.io/otel/exporters/prometheus` to properly handle
unit suffixes when the unit is in brackets.
  E.g. `{spans}`. (#7044)