Add security insights document to repository  (#7129)
Resolve #6245
Resolve #6246
Resolve #6241

Add a "v1.0.0" security insights policy to the repository.

This does not add a "v2.0.0" version of the insights policy as the [CLO
monitor
documentation](https://clomonitor.io/docs/topics/checks/#security-insights)
still reference the "v1.0.0" policy. The policy can be updated if the
CLO tooling is also updated in the future.